% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo_fun.R
\name{elbo_fun}
\alias{elbo_fun}
\title{A function to calculate the expected log sum exponential term for calculation of the elbo}
\usage{
elbo_fun(W_i, xi_i, m_i, V_i, mu_g, Sig_g)
}
\arguments{
\item{W_i}{count data the ith observation.}

\item{xi_i}{variational parameter xi, for the ith observation.}

\item{m_i}{variational parameter m, for the ith observation.}

\item{V_i}{variational parameter V, for the ith observation.}

\item{mu_g}{mean parameter for the latent Gaussian variable from the gth component.}

\item{Sig_g}{variance parameter for the latent Gaussian variable from the gth component.}
}
\description{
This function calculate the expected log sum exponential term which will be used in the calculation of the elbo.
}
\examples{
elbo_fun()
}
\keyword{elbo}
