% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_log_sum_expo.R
\name{expected_log_sum_expo}
\alias{expected_log_sum_expo}
\title{A function to calculate the expected log sum exponential term for calculation of the elbo}
\usage{
expected_log_sum_expo(xi_i, m_i, V_i)
}
\arguments{
\item{xi_i}{variational parameter xi, for the ith observation.}

\item{m_i}{variational parameter m, for the ith observation.}

\item{V_i}{variational parameter V, for the ith observation.}
}
\description{
This function calculate the expected log sum exponential term which will be used in the calculation of the elbo.
}
\examples{
expected_log_sum_expo()
}
\keyword{sum_expo}
