% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_fun.R
\name{initialize_fun}
\alias{initialize_fun}
\title{A function for initialization.}
\usage{
initialize_fun(W, G, K, n, method, nrep = NA, niter = NA, verb = FALSE)
}
\arguments{
\item{W}{observation data: the count data, n*(K+1) matrix where K is the number of taxa.}

\item{G}{number of component.}

\item{K}{number of taxa.}

\item{n}{number of observations.}

\item{method}{string variable, could be "kmeans", "random", and "small_EM".}

\item{nrep}{number of random starts. Only useful if method = "small_EM" is specified. default is 50.}

\item{niter}{number of iterations for each small EM random start. Only useful if method = "small_EM" is specified. default is 10.}

\item{verb}{logical variable. if is TRUE, each iteration of the small_EM start will be printed.}
}
\description{
This function performs initialization of the class membership providing the option of using k-means start, ramdom start, and a small EM or gridsearch type of start: initial randomly on a given number of random starting points, and run for a given number of iterations, then pick the one that ends with highest approximated log-likelihood.
}
\examples{
initialize_fun()
}
\keyword{initialize}
