% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNMVGA.clust.R
\name{LNMVGA.clust}
\alias{LNMVGA.clust}
\title{Function for the Clustering Algorithm}
\usage{
LNMVGA.clust(
  data,
  G,
  initial,
  runtime,
  true_lab = rep(NA, nrow(data)),
  threshold,
  verb,
  maxiter = NA,
  nrep = NA,
  niter = NA
)
}
\arguments{
\item{data}{Matrix of integers. Input data here, data are counts.}

\item{G}{Input a vector specifying all number of components that the model is fit for. E.g., G = c(1:4) will fit models for G = 1,  2, 3, 4.}

\item{initial}{Specify method for initializing z_ig. Possible values could be "kmeans", "random", "small_EM". Default is "kmeans".}

\item{runtime}{Logical variable, if outputting the running time of the whole procedure or not.}

\item{true_lab}{Vector of true class membership. If not specified, defalt is NA for all observations.}

\item{threshold}{Threshold for the Atiken's stopping creterion for convergence.}

\item{verb}{Logical variable, if the key steps of the algortihm and approximated loglikelihood for each iteration are printed.}

\item{maxiter}{Maximum number of iteration. If specified, algorithm will stop by either below the threshold or maxiter reached. If not specified, algorithm will only be monitored by convergence criterion.}

\item{nrep}{Default is NA. Only needed if "small_EM" is specified for initial. Number of random starts for the small EM initialization.}

\item{niter}{Default is NA. Only needed if "small_EM" is specified for initial. Number of iterations for each random start in the small EM initialization.}
}
\value{
A list contains the models selected by BIC or ICL, classification results, BIC/ICL values for all fitted models, and parameter estimation for the selcted model as a vector. Specicially, result.BIC or result.ICL outputs a list contains results for model selected by BIC or ICL: class=class membership, zhat= soft class membership posterior probability, mu=estimated mean parameter for the latent variable, Sigma=estimated variance paprameter for the latent variable, G=number of component selected by BIC, BIC=BIC for all G fitted. est.vec.BIC or est.vec.ICL output a vector contains estimated vectorized mean and variance parameters for the latent variable, and number of component selected. runtime Only output if "runtime == TRUE" in the input, the time for finishing the clustering algorithm for all observations fitting models 1:G.
}
\description{
This function is the main clustering function for the proposed algorithm.
}
\examples{
# generate data using Data.temp <- generate_data(G = 2, num_observation = c(50,50), K = 2, true_mu = list(c(0,1,0),c(-2,-5,0)),true_Sig=list(rbind(cbind(diag(1,2),0),0),rbind(cbind(diag(1,2),0),0)), seed.no = 1234, M = 10000, truelab = TRUE)

LNMVGA.clust(data=Data.temp$W,G=c(1:5),initial="small_EM",runtime=TRUE,true_lab=Data.temp$true_lab,threshold=1e-4,nrep=30,niter=50)
}
\keyword{clust}
