% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{A Map Function}
\usage{
map(x)
}
\arguments{
\item{x}{Input of the function is the vector of class membership posterior probabilities}
}
\description{
This function use given posterior probabilities of observations belonging to a specific cluster (columns corresponding to clusters) to generate a vector of component labels.
}
\examples{
map()
}
\keyword{which.max}
