% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_fun.jacobian.R
\name{mk_fun.jacobian}
\alias{mk_fun.jacobian}
\title{A function for newton update of m
This function calculate the jacobian matrix of the elbo with respect to m for it's kth element.}
\usage{
mk_fun.jacobian(x, Sig, vsquare, M, xi)
}
\arguments{
\item{x}{kth element of m, variable of this function}

\item{Sig}{variance parameter for the latent Gaussian variable.}

\item{vsquare}{diagonal elements of the variational parameter V.}

\item{M}{sum of counts in K taxa for one observation.}

\item{xi}{variational parameter xi, vector of n.}
}
\description{
A function for newton update of m
This function calculate the jacobian matrix of the elbo with respect to m for it's kth element.
}
\examples{
mk_fun.jacobian()
}
\keyword{jacobian}
