% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton_up_m.R
\name{newton_up_m}
\alias{newton_up_m}
\title{A function for newton update of m
This function performs newton update of m within one iteration.}
\usage{
newton_up_m(xold, W_i, iSig, mu, vsquare, M, xi, K)
}
\arguments{
\item{xold}{kth element of m, value from last iteration.}

\item{W_i}{ith observation.}

\item{iSig}{(generalized) inverse matrices variance parameter for the latent Gaussian variable.}

\item{mu}{mean parameter for the latent Gaussian variable.}

\item{vsquare}{diagonal elements of the variational parameter V.}

\item{M}{sum of counts in K taxa for one observation.}

\item{xi}{variational parameter xi, vector of n.}

\item{K}{number of taxa.}
}
\description{
A function for newton update of m
This function performs newton update of m within one iteration.
}
\examples{
newton_up_m()
}
\keyword{newton_up_m}
