% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton_up_v.R
\name{newton_up_v}
\alias{newton_up_v}
\title{A function for newton update of V
This function performs newton update of V within one iteration.}
\usage{
newton_up_v(xold, Sig_kk, m_k, M, xi)
}
\arguments{
\item{xold}{kth diagonal element of V, value from last iteration.}

\item{Sig_kk}{\link{k,k}th element of the variance parameter for the latent Gaussian variable,.}

\item{m_k}{kth element of variational parameter m.}

\item{M}{sum of counts in K taxa for one observation.}

\item{xi}{variational parameter xi, vector of n.}
}
\description{
A function for newton update of V
This function performs newton update of V within one iteration.
}
\examples{
newton_up_v()
}
\keyword{vk}
