% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmap.R
\name{unmap}
\alias{unmap}
\title{An  Unmap Function}
\usage{
unmap(mapz, G)
}
\arguments{
\item{mapz}{Vector of class membership}

\item{G}{Number of component in assumed}
}
\value{
A matrix with the \link{i,j}th entry equals 1 when the ith observation belongs to the jth group.
}
\description{
This function does the opposite to the map() function. It takes the class membership and unmap it to a matrix with entries that equal 1 if the component membership corresponded to that column. It will not drop empty column/class.
}
\examples{
unmap()
}
\keyword{mapz}
