% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varpar_fun.R
\name{varpar_fun}
\alias{varpar_fun}
\title{A function to update the variational parameters.}
\usage{
varpar_fun(W, m, V, mu, Sig, K, z)
}
\arguments{
\item{W}{observation data: the count data, n*(K+1) matrix where K is the number of taxa.}

\item{m}{variational parameter m from last iteration.}

\item{V}{variational parameter Vfrom last iteration.}

\item{mu}{mean parameter for the latent Gaussian variable, list of g vectors of (K+1).}

\item{Sig}{variance parameter for the latent Gaussian variable, list of g (K+1)*(K+1) matrices.}

\item{K}{number of taxa.}

\item{z}{class membership matrix at the current iteration.}
}
\description{
This function use the newton update realted function and perform update on the variational parameters in one iteration.
}
\examples{
varpar_fun()
}
\keyword{varpar}
