% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_fun.R
\name{z_fun}
\alias{z_fun}
\title{A function to calculate posterior clustering probabilities}
\usage{
z_fun(W, xi, m, V, mu, Sig, pi_g, G)
}
\arguments{
\item{W}{observation data: the count data, n*(K+1) matrix where K is the number of taxa.}

\item{xi}{variational parameter xi, vector of n.}

\item{m}{variational parameter m, vector of n.}

\item{V}{variational parameter V, list of n matrices, each element is a (K+1)*(K+1) matrix.}

\item{mu}{mean parameter for the latent Gaussian variable, list of g vectors of (K+1).}

\item{Sig}{variance parameter for the latent Gaussian variable, list of g (K+1)*(K+1) matrices.}

\item{pi_g}{proportion vector of each component.}

\item{G}{number of component.}
}
\description{
This function calculate z, a matrix with each row representing the probability of the observations come from the corresponding group.
}
\examples{
z_fun()
}
\keyword{z_clus}
