% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo_fun.R
\name{elbo_fun}
\alias{elbo_fun}
\title{An internal function to calculate the expected log sum exponential term for calculation of the elbo
This function takes in the parameters for all groups and output a matrix, with the element from the ith row and gth column representing the elbo calculated for the ith observation under the condition that it comes from the gth group.}
\usage{
elbo_fun(W, m, Vmat, V, mu, red_sig, Sig, G)
}
\arguments{
\item{W}{count data}

\item{m}{variational parameter m}

\item{Vmat}{internal parameter, a matrix of n rows as n being number of observations; each row is the diagonal element of the variational parameter V.}

\item{V}{variational parameter V, a list of n matrices, n being number of observations.}

\item{mu}{mean parameter for the latent Gaussian variable.}

\item{red_sig}{internal parameter, a list of G K*K dimensional matrices from the variance parameter.}

\item{Sig}{variance parameter for the latent Gaussian variable, a list of G (K+1)*(K+1) matrices, G is the numbe of component and K+1 is the number of taxa (dimension of the count data).}

\item{G}{number of components.}
}
\description{
An internal function to calculate the expected log sum exponential term for calculation of the elbo
This function takes in the parameters for all groups and output a matrix, with the element from the ith row and gth column representing the elbo calculated for the ith observation under the condition that it comes from the gth group.
}
\examples{
elbo_fun()
}
\keyword{elbo}
