% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.approx.R
\name{loglik.approx}
\alias{loglik.approx}
\title{A function to calculate the approximated log-likelihood}
\usage{
loglik.approx(W, xi, m, V, mu, Sig, pi_g, G)
}
\arguments{
\item{W}{observation data: the count data, n*(K+1) matrix where K is the number of taxa.}

\item{xi}{variational parameter xi, vector of n.}

\item{m}{variational parameter m, vector of n.}

\item{V}{variational parameter V, list of n matrices, each element is a (K+1)*(K+1) matrix.}

\item{mu}{mean parameter for the latent Gaussian variable, list of g vectors of (K+1).}

\item{Sig}{variance parameter for the latent Gaussian variable, list of g (K+1)*(K+1) matrices.}

\item{pi_g}{proportion vector of each component.}

\item{G}{number of component.}
}
\description{
This function calculate approximated log-likelihood using the elbo.
}
\examples{
loglik.approx()
}
\keyword{loglik}
