% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vk_fun.prime.R
\name{vk_fun.prime}
\alias{vk_fun.prime}
\title{Another function for newton update of V
This function calculate the second derivative of the elbo with respect to V for it's kth diagonal element.}
\usage{
vk_fun.prime(x, iSig_kk, m_k, M, xi)
}
\arguments{
\item{x}{kth diagonal element of V, variable of this function}

\item{iSig_kk}{k-th diagonal element of the (generalized) inverse variance parameter for the latent Gaussian variable.}

\item{m_k}{kth element of variational parameter m.}

\item{M}{sum of counts in K taxa for one observation.}

\item{xi}{variational parameter xi, vector of n.}
}
\description{
Another function for newton update of V
This function calculate the second derivative of the elbo with respect to V for it's kth diagonal element.
}
\examples{
vk_fun.prime()
}
\keyword{prime}
