% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vk_fun.R
\name{vk_fun}
\alias{vk_fun}
\title{A function for newton update of V
This function calculate the first derivative of the elbo with respect to V for it's kth diagonal element.}
\usage{
vk_fun(x, Sig_kk, m_k, M, xi)
}
\arguments{
\item{x}{kth diagonal element of V, variable of this function}

\item{Sig_kk}{\link{k,k}th element of the variance parameter for the latent Gaussian variable,.}

\item{m_k}{kth element of variational parameter m.}

\item{M}{sum of counts in K taxa for one observation.}

\item{xi}{variational parameter xi, vector of n.}
}
\description{
A function for newton update of V
This function calculate the first derivative of the elbo with respect to V for it's kth diagonal element.
}
\examples{
vk_fun()
}
\keyword{vk}
